Allow escaping of references and placeholders
JIRA: LMCROSSITXSADEPLOY-1586
