+++ com.sap.cloud.lm.sl.cf.core/src/test/resources/com/sap/cloud/lm/sl/cf/core/helpers/mtad-properties-resolver-test/mtad-with-route-result.json
